pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anse"
version = "0.1.0"
description = "Agent Nervous System Engine - Toolbox for AI agents with sensors and actuators"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "websockets>=10.0",
    "opencv-python>=4.5",
    "sounddevice>=0.4",
    "soundfile>=0.11",
    "pyttsx3>=2.90",
    "pyyaml>=6.0",
    "numpy>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=22.0",
    "mypy>=0.950",
]

[tool.setuptools]
packages = ["anse", "anse.tools", "anse.safety", "anse.examples"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py38']
